% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityIndexFunctions.R
\name{getMobilityIndices}
\alias{getMobilityIndices}
\title{Calculates Mobility Indices}
\usage{
getMobilityIndices(dat, col_x, col_y, type, indices = "all", num_ranks,
  exclude_value, bounds, strict = TRUE)
}
\arguments{
\item{strict}{}
}
\description{
Calculates Mobility Indices
}
