% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityMatrixFunctions.R
\name{getTMatrix}
\alias{getTMatrix}
\title{Calculates Transition Matrix}
\usage{
getTMatrix(dat, col_x, col_y, type, probs = TRUE, num_ranks,
  exclude_value, bounds, strict = TRUE)
}
\arguments{
\item{dat}{a dataframe in the mobilityIndexR schema}

\item{col_x}{a character string denoting the first column to be used in the transition matrix}

\item{col_y}{a character string denoting the second column to be used in the transition matrix}

\item{type}{a character string indicating the type of transition matrix;
accepts relative, mixed, and absolute.}

\item{probs}{logical. If TRUE, values in transition matrix are probabilities;
if FALSE, values in transition matrix are counts}

\item{num_ranks}{an integer specifying the number of ranks for a relative or mixed transition matrix}

\item{exclude_value}{a single numeric value to relegate to an exclusive bottom rank in the transition matrix}

\item{bounds}{a sequence of numeric bounds for defining absolute transition matrix ranks}

\item{strict}{logical.}
}
\value{
Returns transition matrix as a Matrix
}
\description{
Returns transition matrix from two columns in dataset. Supports relative, mixed,
and absolute transition matrices.
}
