% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{gradeMobility}
\alias{gradeMobility}
\title{Example Grades Dataset}
\format{
A \code{data.frame} with 13 columns:
\describe{
 \item{id}{A unique integer id for each row representing a student}
 \item{cohort}{Class of student}
 \item{t0}{Grade at time 0}
 \item{t1}{Grade at time 1}
 \item{t2}{Grade at time 2}
 \item{t3}{Grade at time 3}
 \item{t4}{Grade at time 4}
 \item{t5}{Grade at time 5}
 \item{t6}{Grade at time 6}
 \item{t7}{Grade at time 7}
 \item{t8}{Grade at time 8}
 \item{t9}{Grade at time 9}
 \item{t10}{Grade at time 10}
 }
}
\usage{
gradeMobility
}
\description{
This is an example grade dataset for measuring grade mobility.
}
\keyword{datasets}
