% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{incomeZeroInfMobility}
\alias{incomeZeroInfMobility}
\title{Example Zero-Inflated Income Dataset}
\format{
A \code{data.frame} with 13 columns:
\describe{
 \item{id}{A unique integer id for each row representing an individual}
 \item{cohort}{Year when first income is observed}
 \item{t0}{Income at time 0}
 \item{t1}{Income at time 1}
 \item{t2}{Income at time 2}
 \item{t3}{Income at time 3}
 \item{t4}{Income at time 4}
 \item{t5}{Income at time 5}
 \item{t6}{Income at time 6}
 \item{t7}{Income at time 7}
 \item{t8}{Income at time 8}
 \item{t9}{Income at time 9}
 \item{t10}{Income at time 10}
}
}
\usage{
incomeZeroInfMobility
}
\description{
These is an example income dataset to measure income mobility with an inflated number of zeros.
Zero-inflated data is often observed in practice and will cause difficulties for some transition matrix and mobility
index approaches.
}
\keyword{datasets}
