% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityIndexR.R
\docType{package}
\name{mobilityIndexR}
\alias{mobilityIndexR}
\alias{mobilityIndexR-package}
\title{mobilityIndexR: A package for calculating transition matrices and mobility indices}
\description{
Blah blah blah
}
\section{mobilityIndexR functions}{

getTMatrix
}

