% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityIndexR.R
\docType{package}
\name{mobilityIndexR}
\alias{mobilityIndexR}
\title{mobilityIndexR: A package for calculating transition matrices and mobility indices}
\description{
This package measures mobility in a population through transition matrices and mobility indices.
Relative, mixed, and absolute transition matrices are supported for various use cases. With respect to indices,
the Prais-Bibby, Absolute Movement, Origin Specific, and Shorrocks indices are supported. Example income and
grade data are included for demos.
}
\section{mobilityIndexR functions}{

getTMatrix
getMobilityIndices
getHypothesisTest
}

